Remove core directory - move files to /pipeline
